+++ Program.cs
@@ -2,7 +2,8 @@
 global using ZeroElectric.Vinculum;
 using System.Numerics;
 using BuildingGame;
-using BuildingGame.IO;
+using BuildingGame.Tiles;
+using BuildingGame.Tiles.IO;
 
 SetConfigFlags(ConfigFlags.FLAG_VSYNC_HINT);
 SetConfigFlags(ConfigFlags.FLAG_WINDOW_RESIZABLE);
@@ -13,8 +14,8 @@ Image icon = LoadImage("Assets/Icon.png");
 SetWindowIcon(icon);
 UnloadImage(icon);
 
-BGWorld2IO.Register();
-LvlIO.Register();
+BGWorld2Format.Register();
+LvlFormat.Register();
 
 World world = new World(256, 256);
 world.Load();
