+++ Tiles/Data/TileInfo.cs
@@ -1,4 +1,4 @@
-namespace BuildingGame;
+namespace BuildingGame.Tiles.Data;
 
 public record TileInfo(byte Id, TileFlags Flags)
 {
